fix: persist app-host process authority
